back/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
cryptography==43.0.3
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==0.61.1
django-appconf==1.1.0
django-cors-headers==4.6.0
django-debug-toolbar==4.3.0
django-imagekit==5.0.0
djangorestframework==3.15.2
dotenv==0.9.9
drf-yasg==1.21.7
exceptiongroup==1.3.0
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.10.0
joblib==1.5.1
numpy==2.0.2
oauthlib==3.2.2
openai==1.82.0
packaging==25.0
pilkit==3.0
pillow==10.3.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
scikit-learn==1.6.1
scipy==1.13.1
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
Wikipedia-API==0.8.1
